Implement Painter class

+++ include/event/mouse_button_event.hpp
@@ -30,7 +30,7 @@ namespace cozz {
 
 class MouseButtonEvent final : public Event {
   public:
-    MouseButtonEvent(uint32_t window_id, KeyMap button, bool pressed, uint8_t clicks, int32_t x, int32_t y);
+    MouseButtonEvent(uint32_t window_id, KeyMap button, bool pressed, uint8_t clicks, uint64_t x, uint64_t y);
 
     KeyMap GetButton() const;
 
@@ -38,14 +38,14 @@ class MouseButtonEvent final : public Event {
 
     uint8_t GetClicksCount() const;
 
-    std::pair<int32_t, int32_t> GetPosition() const;
+    std::pair<uint64_t, uint64_t> GetPosition() const;
 
   protected:
     KeyMap button_;
     bool pressed_;
     uint8_t clicks_;
-    int32_t x_;
-    int32_t y_;
+    uint64_t x_;
+    uint64_t y_;
 };
 
 }  // namespace cozz
