+++ include/event/mouse_motion_event.hpp
@@ -37,9 +37,9 @@ class MouseMotionEvent final : public Event {
         kExtra2Pressed = 0x10, /* 0001 0000 */
     };
 
-    MouseMotionEvent(uint32_t window_id, ButtonState state, int32_t x, int32_t y);
+    MouseMotionEvent(uint32_t window_id, ButtonState state, uint64_t x, uint64_t y);
 
-    std::pair<int32_t, int32_t> GetPosition() const;
+    std::pair<uint64_t, uint64_t> GetPosition() const;
 
     bool IsLeftButtonPressed() const;
 
@@ -53,8 +53,8 @@ class MouseMotionEvent final : public Event {
 
   protected:
     ButtonState state_;
-    int32_t x_;
-    int32_t y_;
+    uint64_t x_;
+    uint64_t y_;
 };
 
 MouseMotionEvent::ButtonState operator|(MouseMotionEvent::ButtonState a, MouseMotionEvent::ButtonState b);
