+++ source/event/window_resized_event.cpp
@@ -21,9 +21,9 @@
 
 namespace cozz {
 
-WindowResizedEvent::WindowResizedEvent(uint32_t window_id, int32_t new_width, int32_t new_height)
+WindowResizedEvent::WindowResizedEvent(uint32_t window_id, uint64_t new_width, uint64_t new_height)
     : WindowEvent(window_id), new_width_(new_width), new_height_(new_height) {}
 
-std::pair<int32_t, int32_t> WindowResizedEvent::GetSize() const { return std::make_pair(new_width_, new_height_); }
+std::pair<uint64_t, uint64_t> WindowResizedEvent::GetSize() const { return std::make_pair(new_width_, new_height_); }
 
 }  // namespace cozz
